requirements.txt
django-multiselectfield==0.1.8
Pillow==5.2.0
django==2.0.2
django-ckeditor==5.6.1
django-autocomplete-light==3.3.0
django-querysetsequence==0.8